+++ src/main/java/com/project/snsserver/domain/member/controller/MemberController.java
@@ -6,10 +6,7 @@ import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -20,21 +17,16 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.project.snsserver.domain.member.model.dto.request.LoginRequest;
-import com.project.snsserver.domain.member.model.dto.response.LoginResponse;
 import com.project.snsserver.domain.member.model.dto.request.LogoutRequest;
-import com.project.snsserver.domain.member.model.dto.response.MemberDetailResponse;
-import com.project.snsserver.domain.member.model.dto.response.ProfileResponse;
-import com.project.snsserver.domain.member.model.dto.response.ReissueTokenResponse;
 import com.project.snsserver.domain.member.model.dto.request.SendAuthCodeRequest;
 import com.project.snsserver.domain.member.model.dto.request.SignUpRequest;
-import com.project.snsserver.domain.member.model.dto.response.SignUpResponse;
-import com.project.snsserver.domain.member.model.dto.request.UpdateNicknameRequest;
-import com.project.snsserver.domain.member.model.dto.request.UpdatePasswordRequest;
 import com.project.snsserver.domain.member.model.dto.request.VerifyAuthCodeRequest;
-import com.project.snsserver.domain.member.model.dto.request.WithdrawRequest;
+import com.project.snsserver.domain.member.model.dto.response.LoginResponse;
+import com.project.snsserver.domain.member.model.dto.response.ProfileResponse;
+import com.project.snsserver.domain.member.model.dto.response.ReissueTokenResponse;
+import com.project.snsserver.domain.member.model.dto.response.SignUpResponse;
 import com.project.snsserver.domain.member.model.entity.Member;
 import com.project.snsserver.domain.member.service.MemberService;
-import com.project.snsserver.domain.security.CustomUserDetails;
 import com.project.snsserver.global.util.AuthMember;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -136,72 +128,6 @@ public class MemberController {
 	public ResponseEntity<Map<String, String>> logout(@RequestBody LogoutRequest request, @AuthMember Member member) {
 
 		Map<String, String> response = memberService.logout(request, member.getEmail());
-		return ResponseEntity.ok(response);
-	}
-
-	/**
-	 * 회원 비밀번호 변경
-	 */
-	@Operation(summary = "회원 비밀번호 변경")
-	@PatchMapping("/info/password")
-	public ResponseEntity<Map<String, String>> updatePassword(@RequestBody @Valid UpdatePasswordRequest request,
-		@AuthenticationPrincipal CustomUserDetails userDetails) {
-
-		Map<String, String> response = memberService.updatePassword(request, userDetails.getUsername());
-		return ResponseEntity.ok(response);
-	}
-
-	/**
-	 * 회원 닉네임 변경
-	 */
-	@Operation(summary = "회원 닉네임 변경")
-	@PatchMapping("/info/nickname")
-	public ResponseEntity<Map<String, String>> updateNickname(@RequestBody @Valid UpdateNicknameRequest request,
-		@AuthenticationPrincipal CustomUserDetails userDetails) {
-
-		Map<String, String> response = memberService.updateNickname(request, userDetails.getUsername());
-		return ResponseEntity.ok(response);
-	}
-
-	/**
-	 * 회원 프로필 이미지 변경
-	 */
-	@Operation(summary = "회원 프로필 이미지 변경")
-	@PatchMapping("/info/profile")
-	public ResponseEntity<Map<String, String>> updateProfileImg(@RequestPart(value = "image") MultipartFile file,
-		@AuthenticationPrincipal CustomUserDetails userDetails) {
-		Map<String, String> response = memberService.updateProfileImg(file, userDetails.getUsername());
-		return ResponseEntity.ok(response);
-	}
-
-	/**
-	 * 회원 탈퇴
-	 */
-	@Operation(summary = "회원 탈퇴")
-	@DeleteMapping("/info")
-	public ResponseEntity<Map<String, String>> withdraw(@RequestBody @Valid WithdrawRequest request,
-		@AuthenticationPrincipal CustomUserDetails userDetails) {
-
-		Map<String, String> response = memberService.withdraw(request, userDetails.getUsername());
-		return ResponseEntity.ok(response);
-	}
-
-	/**
-	 * 회원 정보 조회
-	 */
-	@Operation(summary = "나의 회원 상세 정보 조회")
-	@GetMapping("/info")
-	public ResponseEntity<MemberDetailResponse> getMyMemberDetail(@AuthMember Member member) {
-
-		MemberDetailResponse response = memberService.getMyMemberDetail(member);
-		return ResponseEntity.ok(response);
-	}
-
-	@Operation(summary = "다른 회원 상세 정보 조회")
-	@GetMapping("/{memberId}/info")
-	public ResponseEntity<MemberDetailResponse> getMemberDetail(@PathVariable Long memberId) {
-
-		MemberDetailResponse response = memberService.getMemberDetail(memberId);
 		return ResponseEntity.ok(response);
 	}
