+++ src/main/java/com/project/snsserver/domain/member/service/MemberService.java
@@ -1,22 +1,18 @@
 package com.project.snsserver.domain.member.service;
 
+import java.util.Map;
+
+import org.springframework.web.multipart.MultipartFile;
+
 import com.project.snsserver.domain.member.model.dto.request.LoginRequest;
 import com.project.snsserver.domain.member.model.dto.request.LogoutRequest;
 import com.project.snsserver.domain.member.model.dto.request.SignUpRequest;
-import com.project.snsserver.domain.member.model.dto.request.UpdateNicknameRequest;
-import com.project.snsserver.domain.member.model.dto.request.UpdatePasswordRequest;
 import com.project.snsserver.domain.member.model.dto.request.VerifyAuthCodeRequest;
-import com.project.snsserver.domain.member.model.dto.request.WithdrawRequest;
 import com.project.snsserver.domain.member.model.dto.response.LoginResponse;
-import com.project.snsserver.domain.member.model.dto.response.MemberDetailResponse;
 import com.project.snsserver.domain.member.model.dto.response.ProfileResponse;
 import com.project.snsserver.domain.member.model.dto.response.ReissueTokenResponse;
 import com.project.snsserver.domain.member.model.dto.response.SignUpResponse;
 import com.project.snsserver.domain.member.model.entity.Member;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.Map;
 
 public interface MemberService {
 
@@ -59,36 +55,6 @@ public interface MemberService {
 	 * 회원 로그아웃
 	 */
 	Map<String, String> logout(LogoutRequest request, String email);
-
-	/**
-	 * 회원 비밀번호 수정
-	 */
-	Map<String, String> updatePassword(UpdatePasswordRequest request, String email);
-
-	/**
-	 * 회원 닉네임 수정
-	 */
-	Map<String, String> updateNickname(UpdateNicknameRequest request, String email);
-
-	/**
-	 * 회원 프로필 이미지 수정
-	 */
-	Map<String, String> updateProfileImg(MultipartFile file, String email);
-
-	/**
-	 * 회원 탈퇴
-	 */
-	Map<String, String> withdraw(WithdrawRequest request, String email);
-
-	/**
-	 * 다른 회원 정보 상세 조회
-	 */
-	MemberDetailResponse getMemberDetail(Long memberId);
-
-	/**
-	 * 회원 정보 상세 조회
-	 */
-	MemberDetailResponse getMyMemberDetail(Member member);
 
 	/**
 	 * 회원 프로필 조회
